Refactored
Refactored Form and Dialog
