+++ will-call/frontend/src/components/Admin/SubmitDiaologBox.js
@@ -1,39 +1,19 @@
-// import React, { useState } from 'react'
-// import { Button, Dialog, DialogTitle, DialogActions } from '@mui/material'
+import React from 'react'
+import { Button, Dialog, DialogTitle, DialogActions } from '@mui/material';
 
-// export default function SubmitDiaologBox(handleDeleteCall, ticket) {
-//   const [open, setOpen] = useState(false);
 
-//   const handleOpenDialog = () => {
-//     setOpen(true);
-//   };
-
-//   const handleCloseDialog = () => {
-//     setOpen(false);
-//   };
-
-//   const handleSubmit = () => {
-//     // Handle submit logic here
-//     // This function will be called when the user confirms
-//     // Close the dialog
-//     handleCloseDialog();
-//   };
-//   return (
-//     <div>
-//       <Button variant="contained" color="primary" onClick={handleOpenDialog}>
-//         Submit
-//       </Button>
-//       <Dialog open={open} onClose={handleCloseDialog}>
-//         <DialogTitle>Are you sure you want to submit?</DialogTitle>
-//         <DialogActions>
-//           <Button onClick={handleCloseDialog} color="primary">
-//             Cancel
-//           </Button>
-//           <Button onClick={handleSubmit} color="primary" autoFocus>
-//             Submit
-//           </Button>
-//         </DialogActions>
-//       </Dialog>
-//     </div>
-//   )
-// }
+export default function SubmitDiaologBox({ open, handleCloseDialog, handleDelete }) {
+  return (
+    <Dialog open={open} onClose={handleCloseDialog} >
+      <DialogTitle>Are you sure you want to DELETE?</DialogTitle>
+      <DialogActions>
+        <Button sx={{ color: 'grey' }} onClick={handleCloseDialog} color="primary">
+          Cancel
+        </Button>
+        <Button sx={{ color: 'salmon' }} onClick={() => handleDelete()} color="primary" autoFocus>
+          Delete
+        </Button>
+      </DialogActions>
+    </Dialog>
+  )
+}
